Openclaw_PwC/requirements.txt
# PWC AI Assistant PoC - Dependencies
# Comprehensive PC control tools with LangChain

# === Core: LangChain & LLM ===
langchain>=0.3.14
langchain-openai>=0.3.0
langgraph>=0.2.60

# === Core: API & Server ===
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# === Core: Environment ===
python-dotenv>=1.0.1

# === MCP Integration (optional) ===
langchain-mcp-adapters>=0.1.0
mcp>=1.3.0
anyio>=4.7.0

# === File & Office Tools ===
openpyxl>=3.1.0           # Excel read/write
python-docx>=1.1.0        # Word read/write
chardet>=5.2.0            # Encoding detection

# === Web Tools ===
requests>=2.31.0          # HTTP requests
beautifulsoup4>=4.12.0    # HTML parsing
duckduckgo-search>=6.0.0  # Web search (no API key)

# === System Tools ===
psutil>=5.9.0             # Process management
pillow>=10.0.0            # Screenshot (Windows)
